Updated tests to reflect changes
The routing container must be specified as an option and so tests have been created to ensure that when it is not, no changes are made to the system. An option has been added to the test which should be correctly create the network which maintains compatability with previous tests.